requirements.txt
backoff==1.11.1
python-dateutil==2.8.2
google-api-python-client==2.40.0
google-cloud-storage==2.2.1
google-cloud-tasks==2.8.1
